1. 스테이지 1 디자인 2. Item Block(충돌 시 아이템 생성) 3. Block(캐릭터가 부술 수 있는 블럭) 4. 벽, Time, Score, Coin Count 추가

+++ WonEngineSource/WInput.cpp
@@ -129,16 +129,16 @@ void Won::WInput::SetMousePosition()
 	GetCursorPos(&mousePos);
 	ScreenToClient(Engine.GetEngineWND(), &mousePos);
 
-	UINT width = Engine.GetWidth();
-	UINT height = Engine.GetHeight();
+	LONG width = Engine.GetWidth();
+	LONG height = Engine.GetHeight();
 
 	mMousePosition.X = -1.0f;
 	mMousePosition.Y = -1.0f;
 
 	if (mousePos.x > 0 && mousePos.x < width)
-		mMousePosition.X = mousePos.x;
+		mMousePosition.X = static_cast<float>(mousePos.x);
 
 	if (mousePos.y > 0 && mousePos.y < height)
-		mMousePosition.Y = mousePos.y;
+		mMousePosition.Y = static_cast<float>(mousePos.y);
 
 }
